Restructure app and use loro text crdt (#41)
Introduce a new `doc` crate that exposes GObject API to be used in the
app crate. This integrates also the work @adzialocha did to replace
automerge with loro.

Closes: https://github.com/p2panda/aardvark/issues/37
Closes: https://github.com/p2panda/aardvark/issues/19

Sorry for stealing @adzialocha changes in #33 and #39 hope that's fine.
Let me know if you want me to split the commit even more.

---------

Co-authored-by: adz <x12@adz.garden>
Co-authored-by: adz <adzialocha@users.noreply.github.com>