Separate networking- and application-layers
Refactor the "backend" part into separate layers: one Network struct
concerning p2p networking, and a Node struct offering the API towards
the application layer, using the network internally. With this it should
be clearer to understand where a "generic" p2p backend lives and where
the data-types for the text editor begin.