Fix: Local Loro commit can contain multiple changes (#54)
This is a minor bug fix of a feature we currently don't use, but it's
annoying enough as it causes the app to crash.

I've assumed that loro will always only generate one "delta" after a
change was made and the document was committed, but this was not
correct. This change will allow multiple deltas.