fix: Topic ids are not always space ids
Space ids are actor ids which again are public keys. We use the gossip
overlay to learn about changes around a topic, which can be both a space
id or the network-wide gossip overlay. The latter is not a public key,
this is why the conversion fails (due to an invalid edwards point).