node: Track Authors online state with Hello/Bye messages
This implements a custom mechanism to track whether authors are online.
Each author sends a `Hello` message when they join a document, all other
authors respond with a `Ping` message. Before a author leaves a document
they send a `Bye` message. Authors also send a `Ping` message every 30
seconds to signal that they are still online, since it may be possible
that an author is unable to send a `Bye` message. If no `Ping` is
received in 60 seconds then it's assumed the author went offline.

Closes: https://github.com/p2panda/reflection/issues/70