node: Add OperationStore struct
This adds an abstraction over the SqliteStore from p2panda. The main
purpose is to guarantee sequential writes which SqliteStore doesn't.
Eventually we could also cache inside this struct the latest operation
for each subscribed document, so that we don't need to read it from
sqlite on every new operation.