node: Return Subscription struct when subscribing to a document
We want users only to interact with a document once they have subscribed
to it e.g. we need to prevent users from sending operations before they
subscribed. By only using the `Subscription` struct to handle interaction with
the document we can guarantee this programmatically. The current methods
on `Node` to interact with a document will be removed in a future
commit.